+++ qt_widgets/DNRAnalogClock.h
@@ -57,7 +57,7 @@ class QDESIGNER_WIDGET_EXPORT DNRAnalogClock : public QWidget
     Q_PROPERTY(int EndTimeMinute READ getEndTimeMinute WRITE setEndTimeMinute);
     Q_PROPERTY(int EndTimeSecond READ getEndTimeSecond WRITE setEndTimeSecond);
     Q_PROPERTY(QColor EndTimeColor READ getEndTimeColor WRITE setEndTimeColor);
-    Q_PROPERTY(int CountDownTime READ getCountDownTime WRITE setCountDownTime);
+    Q_PROPERTY(float CountDownTime READ getCountDownTime WRITE setCountDownTime);
     Q_PROPERTY(int CountDownWidth READ getCountDownWidth WRITE setCountDownWidth);
     Q_PROPERTY(int CountDownSpacing READ getCountDownSpacing WRITE setCountDownSpacing);
     Q_PROPERTY(QColor CountDownColor READ getCountDownColor WRITE setCountDownColor);
@@ -172,8 +172,8 @@ public:
     void setEndTimeColor(QColor NewEndTimeColor);
     QColor getEndTimeColor();
 
-    void setCountDownTime(int NewCountDownTime);
-    int getCountDownTime();
+    void setCountDownTime(float NewCountDownTime);
+    float getCountDownTime();
 
     void setCountDownWidth(int NewCountDownWidth);
     int getCountDownWidth();
@@ -196,7 +196,7 @@ protected:
     int FEndTimeSecond;
     QColor FEndTimeColor;
 
-    int FCountDownTime;
+    float FCountDownTime;
     int FCountDownWidth;
     int FCountDownSpacing;
     QColor FCountDownColor;
